backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
stanza==1.8.2
python-dotenv==1.0.1
httpx==0.27.0
inltk==0.9
requests==2.32.3
